Split main entry file
