+++ snn_hfo_ieeg/run.py
@@ -1,0 +1,16 @@
+from snn_hfo_ieeg.functions.filter import *
+from snn_hfo_ieeg.functions.dynapse_biases import *
+from snn_hfo_ieeg.functions.signal_to_spike import *
+from snn_hfo_ieeg.functions.hfo_detection import *
+from snn_hfo_ieeg.hfo_detection import run_hfo_detection_with_configuration
+from snn_hfo_ieeg.argument_parsing import parse_arguments, convert_arguments_to_config, convert_arguments_to_custom_overrides
+
+
+def run_hfo_detection(hfo_cb):
+    arguments = parse_arguments()
+    configuration = convert_arguments_to_config(arguments)
+    custom_overrides = convert_arguments_to_custom_overrides(arguments)
+    run_hfo_detection_with_configuration(
+        configuration=configuration,
+        custom_overrides=custom_overrides,
+        hfo_cb=hfo_cb)
