+++ snn_hfo_ieeg/stages/all.py
@@ -7,7 +7,7 @@ HFO_DETECTION_STEP_SIZE = 0.01
 HFO_DETECTION_WINDOW_SIZE = 0.05
 
 
-def run_hfo_detection(channel_data, duration, configuration):
+def run_all_hfo_detection_stages(channel_data, duration, configuration):
     filtered_spikes = filter_stage(channel_data, configuration)
     spike_monitor_hidden = snn_stage(
         filtered_spikes=filtered_spikes,
